chore(contracts): move location spawn into create game (#83)

+++ src/events.cairo
@@ -53,8 +53,8 @@ struct Sold {
 struct Traveled {
     game_id: u32,
     player_id: felt252,
-    from_location_id: u32,
-    to_location_id: u32,
+    from_location_id: u8,
+    to_location_id: u8,
 }
 
 #[derive(Drop, Serde)]
